requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
psycopg[binary]==3.1.18
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0